Update Docstrings, revise to add rosbag and subscriber based extraction.
